* Fixup some compiler warnings.
git-archimport-id: falk@cs.fau.de--2008-systemoc-intern/SysteMoC--devel--0.9--patch-110

+++ pkginclude/systemoc/smoc_multiplex_fifo.hpp
@@ -571,7 +571,7 @@ protected:
 
   /// @brief See smoc_port_in_base_if
   size_t inTokenId() const
-    { return -1; }
+    { return static_cast<size_t>(-1); }
 
   /// @brief See smoc_port_in_if
   typename this_type::access_type *getReadPortAccess()
@@ -690,7 +690,7 @@ protected:
 
   /// @brief See smoc_port_in_base_if
   size_t inTokenId() const
-    { return -1; }
+    { return static_cast<size_t>(-1); }
 
   /// @brief See smoc_port_in_if
   AccessImpl *getReadPortAccess()
@@ -809,7 +809,7 @@ protected:
 
   /// @brief See smoc_port_out_base_if
   size_t outTokenId() const
-    { return -1; }
+    { return static_cast<size_t>(-1); }
 
   /// @brief See smoc_port_out_if
   AccessImpl *getWritePortAccess()
